feat: init bottom navigator
